requirements.txt
streamlit
BeautifulSoup4
request
lxml